+++ CHANGELOG.md
@@ -538,6 +538,10 @@ Thanks to the people who made this release happen!
   launched once per changed file, making it possible to use per-file tools like
   `vimdiff` for editing.
 
+* Added support `revsets.git-push` which allows you to customize the default
+  set of commits to push. See issue
+  [#3649](https://github.com/jj-vcs/jj/issues/3650).
+
 ### Fixed bugs
 
 * `jj git remote add` now reports an error instead of panicking when the
@@ -895,6 +899,7 @@ Thanks to the people who made this release happen!
 
 * Conditional configuration based on environment variables with `--when.environments`.
   [#8779](https://github.com/jj-vcs/jj/pull/8779)
+
 
 ### Fixed bugs
 
@@ -1579,6 +1584,7 @@ Thanks to the people who made this release happen!
 * `jj file track` now accepts an `--include-ignored` flag to track files that
   are ignored by `.gitignore` or exceed the `snapshot.max-new-file-size` limit.
   [#2837](https://github.com/jj-vcs/jj/issues/2837)
+
 
 ### Fixed bugs